requirements.txt
Flask==3.1.0
SQLAlchemy==2.0.36
PyMySQL==1.1.1
pandas==2.2.3
APScheduler==3.10.4
python-dotenv==1.0.1
